DMP 1327 shifted api around
