+++ src/main/java/uk/gov/hmcts/darts/hearings/service/HearingsService.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.darts.hearings.service;
 import uk.gov.hmcts.darts.common.entity.HearingEntity;
 import uk.gov.hmcts.darts.hearings.model.EventResponse;
 import uk.gov.hmcts.darts.hearings.model.GetHearingResponse;
-import uk.gov.hmcts.darts.hearings.model.Transcript;
+import uk.gov.hmcts.darts.transcriptions.model.Transcript;
 
 import java.util.List;
