+++ src/main/resources/openapi/cases.yaml
@@ -265,62 +265,6 @@ paths:
               schema:
                 $ref: './problem.yaml'
 
-  /cases/{case_id}/transcripts:
-    get:
-      tags:
-        - Cases
-      summary: Allows the retrieval of transcripts for a given case_id
-      description: |-
-        Retrieves all transcripts for a case_id.
-      parameters:
-        - in: path
-          name: case_id
-          schema:
-            type: integer
-          description: "case_id is the internal cas_id of the case."
-          required: true
-      responses:
-        '200':
-          description: OK
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/transcripts'
-        '401':
-          description: Unauthorised Error
-        '403':
-          description: Forbidden Error
-          content:
-            application/json+problem:
-              schema:
-                allOf:
-                  - $ref: './problem.yaml'
-                  - type: object
-                    required:
-                      - type
-                    properties:
-                      type:
-                        $ref: '#/components/schemas/CasesAuthorisation403ErrorCode'
-              example:
-                type: "AUTHORISATION_100"
-                title: "User is not authorised for the associated courthouse"
-                status: 403
-        '404':
-          description: Not Found Error
-          content:
-            application/json+problem:
-              schema:
-                $ref: './problem.yaml'
-              example:
-                type: "CASE_104"
-                title: "The requested case cannot be found"
-                status: 404
-        '500':
-          description: Internal Server Error
-          content:
-            application/json+problem:
-              schema:
-                $ref: './problem.yaml'
 
   /cases/search:
     get:
@@ -600,36 +544,6 @@ components:
           type: array
           items:
             $ref: '#/components/schemas/judge'
-    transcripts:
-      type: array
-      items:
-        $ref: '#/components/schemas/transcript'
-    transcript:
-      type: object
-      properties:
-        tra_id:
-          type: integer
-          example: 1
-        hea_id:
-          type: integer
-          example: 2
-        hearing_date:
-          type: string
-          format: date
-          example: 2022-05-20
-        type:
-          type: string
-          example: "Sentencing remarks"
-        requested_on:
-          type: string
-          format: date
-          example: 2022-05-22
-        requested_by_name:
-          type: string
-          example: "Joe Bloggs"
-        status:
-          type: string
-          example: "Available"
     defendant:
       type: string
       example: Joe Bloggs
