+++ src/main/resources/openapi/transcriptions.yaml
@@ -157,6 +157,121 @@ paths:
         '401':
           description: Unauthorised Error
 
+
+  /transcripts/hearings/{hearing_id}:
+    get:
+      tags:
+        - Transcription
+      summary: Allows the retrieval of transcripts for a given hearing_id
+      description: |-
+        Retrieves all transcripts for a hearing_id.
+      parameters:
+        - in: path
+          name: hearing_id
+          schema:
+            type: integer
+          description: "hearing_id is the internal hea_id of the case."
+          required: true
+      responses:
+        '200':
+          description: OK
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/transcripts'
+        '401':
+          description: Unauthorised Error
+        '403':
+          description: Forbidden Error
+          content:
+            application/json+problem:
+              schema:
+                allOf:
+                  - $ref: './problem.yaml'
+                  - type: object
+                    required:
+                      - type
+                    properties:
+                      type:
+                        $ref: '#/components/schemas/HearingsAuthorisation403ErrorCode'
+              example:
+                type: "AUTHORISATION_100"
+                title: "User is not authorised for the associated courthouse"
+                status: 403
+        '404':
+          description: Not Found Error
+          content:
+            application/json+problem:
+              schema:
+                $ref: './problem.yaml'
+              example:
+                type: "CASE_104"
+                title: "The requested case cannot be found"
+                status: 404
+        '500':
+          description: Internal Server Error
+          content:
+            application/json+problem:
+              schema:
+                $ref: './problem.yaml'
+  /transcripts/case/{case_id}:
+    get:
+      tags:
+        - Transcription
+      summary: Allows the retrieval of transcripts for a given case_id
+      description: |-
+        Retrieves all transcripts for a case_id.
+      parameters:
+        - in: path
+          name: case_id
+          schema:
+            type: integer
+          description: "case_id is the internal cas_id of the case."
+          required: true
+      responses:
+        '200':
+          description: OK
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/transcripts'
+        '401':
+          description: Unauthorised Error
+        '403':
+          description: Forbidden Error
+          content:
+            application/json+problem:
+              schema:
+                allOf:
+                  - $ref: './problem.yaml'
+                  - type: object
+                    required:
+                      - type
+                    properties:
+                      type:
+                        $ref: '#/components/schemas/CasesAuthorisation403ErrorCode'
+              example:
+                type: "AUTHORISATION_100"
+                title: "User is not authorised for the associated courthouse"
+                status: 403
+        '404':
+          description: Not Found Error
+          content:
+            application/json+problem:
+              schema:
+                $ref: './problem.yaml'
+              example:
+                type: "CASE_104"
+                title: "The requested case cannot be found"
+                status: 404
+        '500':
+          description: Internal Server Error
+          content:
+            application/json+problem:
+              schema:
+                $ref: './problem.yaml'
+
+
 components:
   schemas:
     TranscriptionRequestDetails:
@@ -242,6 +357,46 @@ components:
       enum:
         - "TRANSCRIPTION_105"
       x-enum-varnames: [ TRANSCRIPTION_WORKFLOW_ACTION_INVALID ]
+    transcripts:
+      type: array
+      items:
+        $ref: '#/components/schemas/transcript'
+    transcript:
+      type: object
+      properties:
+        tra_id:
+          type: integer
+          example: 1
+        hea_id:
+          type: integer
+          example: 2
+        hearing_date:
+          type: string
+          format: date
+          example: 2022-05-20
+        type:
+          type: string
+          example: "Sentencing remarks"
+        requested_on:
+          type: string
+          format: date
+          example: 2022-05-22
+        requested_by_name:
+          type: string
+          example: "Joe Bloggs"
+        status:
+          type: string
+          example: "Available"
+    CasesAuthorisation403ErrorCode:
+      type: string
+      enum:
+        - "AUTHORISATION_100"
+      x-enum-varnames: [ USER_NOT_AUTHORISED_FOR_COURTHOUSE ]
+    HearingsAuthorisation403ErrorCode:
+      type: string
+      enum:
+        - "AUTHORISATION_100"
+      x-enum-varnames: [ USER_NOT_AUTHORISED_FOR_COURTHOUSE ]
 
     TranscriptionTypesResponse:
       type: array
